+++ inc/commands/run.js
@@ -32,7 +32,7 @@
     /**
      * The Layer we want to run.
      *
-     * @type Layer|null
+     * @type RunLayer|null
      */
     let theLayer = null;
 
@@ -75,6 +75,10 @@
                     }
 
                     theLayer = application.getLayer(layerName);
+
+                    if (!layerName.runOnly) {
+                        return reject(new Error('The layer ' + layerName + ' for application ' + applicationName + ' is not a runnable layer!'));
+                    }
 
                     resolve();
                 });