+++ CHANGES.md
@@ -3,6 +3,8 @@ This file marks the changes in each version of ADM starting with 0.2.0.
 
 ## 1.2
 ### 1.2.0
+#### Changes
+- Massively speed up the `up` command.
 #### Removed
 - Removed the monitor command since that's not within the scope of the application.