+++ CHANGES.md
@@ -10,6 +10,7 @@ This file marks the changes in each version of ADM starting with 0.2.0.
 - Massively speed up the `up` command.
 #### Fixes
 - Fix issue with the `up` command still attempting to bring an application up when it's all up already.
+- Fix the `run`` command not erroring out when attempting to run a non run layer.
 #### Removed
 - Removed the monitor command since that's not within the scope of the application.